+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
 ## 0.13.0 - TBD
+- Added support for `statistics` schema
+- Upgraded `databento-dbn` to 0.5.1
 - Renamed `booklevel` MBP field to `levels` for brevity and consistent naming
 - Changed `flags` field to an unsigned int
 - Changed default of `ts_out` to `False` for `Live` client
-- Removed `open_interest_qty` and `cleared_volume` fields that were always unset from
-  definition schema
+- Removed `open_interest_qty` and `cleared_volume` fields that were always unset from definition schema
 
 ## 0.12.0 - 2023-05-01
 - Added `Live` client for connecting to Databento's live service